Moved everything into directories. Added a Makefile for each directory that has a .cpp file (not just headers). The Makefile in the src directory runs make in all the subdirectories and links everything together. Moved main.cpp to widget/preview.cpp and made a new main.cpp that doesn't do anything (this will be the main.cpp of the actual game). Running 'make' in the src directory will make both the widget preview and the game (./preview and ./adage). You can 'make preview' or 'make adage' if you want. 'make clean' cleans out the whole directory hierarchy.
To add a .cpp to the makefiles, just add the corresponding .o file in 
that directory's Makefile's OBJECTS line. To add a whole new directory, 
first create a Makefile in the directory (just copy, paste, and edit one 
from another directory) and then add the directory's name in the MODULES 
line of the main Makefile.